moved and refactored SingleOrderControllers
