+++ csc-cmo-ssapp/code/index.html
@@ -50,28 +50,6 @@
       </webc-app-container>
     </webc-app-root>
 
-    <!--<webc-app-root></webc-app-root>-->
-
-    <!--<psk-app-root>-->
-    <!--    <psk-container controller-name="ParticipantController">-->
-    <!--        <psk-condition condition="@identified">-->
-    <!--            <div slot="condition-true">-->
-    <!--                <div class="global_container">-->
-    <!--                    <psk-app-menu hamburger-max-width="0" class="demo">-->
-    <!--                        <psk-label slot="before" class="title" label="@locale.home.title"></psk-label>-->
-    <!--                    </psk-app-menu>-->
-    <!--                    <section>-->
-    <!--                        <psk-app-router failRedirectTo="/Dashboard"></psk-app-router>-->
-    <!--                    </section>-->
-    <!--                </div>-->
-    <!--            </div>-->
-    <!--            <div slot="condition-false">-->
-    <!--                <psk-label class="page-title" label="@locale.home.title"></psk-label>-->
-    <!--            </div>-->
-    <!--        </psk-condition>-->
-    <!--    </psk-container>-->
-    <!--</psk-app-root>-->
-
     <!-- WebCardinal dist -->
     <script type="module" src="webcardinal/webcardinal.js"></script>
   </body>
