+++ csc-cmo-ssapp/code/pages/order.html
@@ -156,7 +156,7 @@
             </div>
           </div>
 
-          <button disabled="@order.couldNotBeReviewed" data-tag="review-order" class="simple-default-button">Review</button>
+          <button disabled="@order.actions.couldNotBeReviewed" data-tag="review-order" class="simple-default-button">Review</button>
 
         </div>
       </div>
