+++ csc-services/controllers/impl/index.js
@@ -1,0 +1,1 @@
+require("./SingleOrderControllerImpl");
