+++ csc-sponsor-ssapp/code/pages/order.html
@@ -351,11 +351,11 @@
           </div>
 
           <div class="row">
-            <button disabled="@order.couldNotBeReviewed" data-tag="review-order" class="col-md-2 simple-default-button">Review</button>
+            <button disabled="@order.actions.couldNotBeReviewed" data-tag="review-order" class="col-md-2 simple-default-button">Review</button>
             <div class="col-md-2"></div>
-            <button disabled="@order.couldNotBeCancelled" data-tag="cancel-order" class="col-md-2 simple-default-button bg-red">Cancel Order</button>
+            <button disabled="@order.actions.couldNotBeCancelled" data-tag="cancel-order" class="col-md-2 simple-default-button bg-red">Cancel Order</button>
             <div class="col-md-2"></div>
-            <button disabled="@order.couldNotBeApproved" data-tag="approve-order" class="col-md-2 simple-default-button bg-green">Approve</button>
+            <button disabled="@order.actions.couldNotBeApproved" data-tag="approve-order" class="col-md-2 simple-default-button bg-green">Approve</button>
           </div>
         </div>
       </div>
